ruleset: use crc32 for reproducibility
When we generate the ID for a rule, we use the builtin hash function to
do so.  Unfortunately, the values output by this function change on each
invocation to prevent hash DoS attacks, meaning our output isn't
consistent, which makes diffs of the output noisy.

It doesn't appear that Google actually cares what the ID of the rule is,
so we don't need to pick anything like a cryptographic hash that would
prevent collisions.  Let's use CRC32, which is consistent and fast, but
not cryptographically secure.  If things change in the future and Google
becomes pickier, we can switch to an encoding of SHA-256.

Since we can no longer pass an arbitrary object, use the repr function
to turn the rule into a string format and then encode it in UTF-8 to
produce the bytes that our hash wants.