Allow fixed timestamps for reproducibility
Currently, we emit the current time into the XML we generate.  This
unfortunately leads to a large amount of diff noise when we compare two
XML files, especially when there are many rules.

Let's offer a way to control the date used so that we can produce nice
diffs.  The SOURCE_DATE_EPOCH environment variable is used when doing
reproducible builds to specify the number of seconds from the Epoch
representing the time to be used during the build.  We can use this here
to allow the user to specify the timestamp to be embedded in their file.
If they don't specify this, we simply use the current time, as before.

Additionally, this means we can add some tests for our new reproducible
behavior that cover the entirety of the XML output.  In the future, we
can verify the results of our ruleset handling in various cases by
generating and comparing the XML.